svg-gen-evaluator/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "svg-gen-evaluator"
version = "0.1.0"
description = "Evaluation toolkit for SVG generation models in Kaggle competitions"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "pandas",
    "torch>=1.10.0",
    "transformers>=4.20.0",
    "cairosvg",
    "Pillow",
    "matplotlib",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black",
    "isort",
    "flake8",
    "pre-commit",
]

[project.scripts]
svg-evaluate = "svg_gen_evaluator.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"